camera: explicitly cast `human_sensing::Landmark::Type`
The new version of clang warns about the use of `enum class`es in
`printf`-like contexts. To avoid build errors, we need to explicitly
`static_cast` it.

BUG=b:305714217
TEST=emerge-jacuzzi cros-camera-libs

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4943252
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>

+++ camera/common/camera_face_detection.cc
@@ -267,7 +267,8 @@ std::string LandmarkTypeToString(human_sensing::Landmark::Type type) {
     case human_sensing::Landmark::Type::kLandmarkUnknown:
       return "Unknown";
   }
-  return base::StringPrintf("Undefined landmark type %d", type);
+  return base::StringPrintf("Undefined landmark type %d",
+                            static_cast<int>(type));
 }
 
 }  // namespace cros
